adding tests for geomag
